refactor: WebChessGame 삭제
WebChessGame 삭제
ChessGame WebChessGame에 있던 reset/createWebBoard 추가하여 클래스 기능 통합
